commit: let --amend=<commit> create an "amend!" commit
Add an optional argument to '--amend' that creates an "amend!" commit
for use with 'rebase -i --autosquash'. If no message is given via
-C/-c/-F/-m then read in the message of <commit> and allow the user to
edit it. If no argument is given the behavior of '--amend' is
unchanged.

I don't think that overloading '--amend' in this way is a
particularly good user interface but unless I can come up with a
better name for the "amend" rebase command doing this is better than
having '--some-other-option' create "amend!" commits.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>