rebase -i: always keep empty amend! commits
Empty amend! commits are used to reword a previous commit when rebasing
so we don't want to discard them. This adds an unfortunate difference of
behaviour between regular rebases which will still discard the empty
amend! commits if the user does not specify --keep-empty and interactive
rebases which will keep them. Unfortunately it is not easy to change the
regular rebase to match the new behavior of interactive rebases.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>