commit.c: add option to create amend! commit
Add amend option to the --fixup to create amend! commit, which helps
to fixup the content and also allows to edit the commit message.

Users-note: Use command `git commit --fixup=amend:<commit>` to
prepare the amend! commit and upon --autosquash it will fixup the
content and replace the commit message with the amend! commit message's
body.

Signed-off-by: Charvi Mendiratta <charvi077@gmail.com>