commit.c: add option to create amend! commit
Add amend option to the --fixup that creates amend! commit and allows
to fixup the content and also to edit the commit message.

Users-note: Use command `git commit --fixup=amend:<commit>` to
prepare the amend! commit and upon --autosquash it will fixup the
content and automatically replace the commit message with the
amend! commit message's body.

Signed-off-by: Charvi Mendiratta <charvi077@gmail.com>