commit.c: implementation of amend! commit
Add amend option to the --fixup to make a amend! commit, which helps
to fixup the content and also allows to edit the commit message.

Users-note: Use command `git commit --fixup=amend:<commit>` to
prepare the amend! commit. And upon --autosquash it will fixup the
content and replace the commit message.