sequencer.c: fixing the bug for the multiple fixup -c chain
Bug : Failed to open the editor if we use sequence `fixup -C`, `fixup -c`,
`fixup -C`. As in this case in function do_pick_commit() as last
command is `fixup -C` and it writes the commit message in
rebase_path_fixup_msg() so it cleans up the message and do not opens the
editor.
So, added the flag check as in this case also we need the editor to be
opened.